+++ src/transforms/mod.rs
@@ -1,0 +1,5 @@
+#![doc = include_str!("../../docs/ngrams.md")]
+
+mod ngrams;
+
+pub use ngrams::{InvalidNgramLength, ngrams};